+++ examples/cpp/quickies/heap.cpp
@@ -7,7 +7,7 @@ void display( char const * msg, std::vector<int> const & v )
 {
   std::cout << msg << ": ";
 
-  for ( int i = 0 ; i < v.size() ; ++i ) {
+  for ( size_t i = 0 ; i < v.size() ; ++i ) {
     std::cout << v[i] << " ";
   }
   std::cout << std::endl;
